+++ hat/api_import/apps.py
@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 
 class VectorConfig(AppConfig):
-    name = "hat.vector_control"
+    name = "hat.api_import"
