+++ iaso/api/common.py
@@ -18,7 +18,7 @@ from rest_framework.response import Response
 from rest_framework.viewsets import ModelViewSet as BaseModelViewSet
 from rest_framework_csv.renderers import CSVRenderer
 
-from hat.vector_control.models import APIImport
+from hat.api_import.models import APIImport
 from iaso.models import OrgUnit, OrgUnitType
 
 logger = logging.getLogger(__name__)
