+++ iaso/management/commands/reimport_failed_imports.py
@@ -1,6 +1,6 @@
 from django.core.management.base import BaseCommand
 
-from hat.vector_control.models import APIImport
+from hat.api_import.models import APIImport
 from iaso.api.instances import import_data as import_instances
 from iaso.api.org_units import import_data as import_units
